requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
pyyaml>=6.0.1
click>=8.1.7
loguru>=0.7.2
joblib>=1.3.0
scipy>=1.11.0